+++ src/plone/app/multilingual/tests/test_selector.py
@@ -15,17 +15,20 @@ from plone.dexterity.utils import createContentInContainer
 from plone.registry.interfaces import IRegistry
 from plone.testing.z2 import Browser
 from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import safe_nativestring
+from six.moves.urllib.parse import parse_qs
+from six.moves.urllib.parse import urlparse
 from zope.component import getUtility
 from zope.event import notify
 from zope.interface import alsoProvides
 from zope.lifecycleevent import ObjectModifiedEvent
 
+import six
 import transaction
 import unittest
 
 
 class EvilObject(object):
-
     def __str__(self):
         raise UnicodeError
 
@@ -33,8 +36,9 @@ class EvilObject(object):
         raise UnicodeError
 
 
-SELECTOR_VIEW_TEMPLATE =\
+SELECTOR_VIEW_TEMPLATE = (
     '%(url)s/@@multilingual-selector/%(tg)s/%(lang)s%(query)s'
+)
 
 
 class TestLanguageSelectorBasics(unittest.TestCase):
@@ -53,35 +57,41 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         self.browser.handleErrors = False
 
     def test_selector_viewlet_is_available(self):
-        selector_viewlet = \
-            LanguageSelectorViewlet(self.portal, self.request, None, None)
+        selector_viewlet = LanguageSelectorViewlet(
+            self.portal, self.request, None, None
+        )
         selector_viewlet.update()
         self.assertTrue(selector_viewlet.available())
 
     def test_selector_viewlet_is_available_but_tool_is_not(self):
-        selector_viewlet = \
-            LanguageSelectorViewlet(self.portal, self.request, None, None)
+        selector_viewlet = LanguageSelectorViewlet(
+            self.portal, self.request, None, None
+        )
         selector_viewlet.update()
         selector_viewlet.tool = None
         self.assertFalse(selector_viewlet.available())
 
     def test_selector_data_not_modified_by_viewlet(self):
-        selector_viewlet = \
-            LanguageSelectorViewlet(self.portal, self.request, None, None)
+        selector_viewlet = LanguageSelectorViewlet(
+            self.portal, self.request, None, None
+        )
         selector_viewlet.update()
         selector_viewlet_languages = selector_viewlet.languages()
 
-        selector_adapter = \
-            LanguageSelector(self.portal, self.request, None, None)
+        selector_adapter = LanguageSelector(
+            self.portal, self.request, None, None
+        )
         selector_adapter.update()
         selector_adapter_languages = selector_adapter.languages()
 
-        self.assertNotEquals(selector_adapter_languages,
-                             selector_viewlet_languages)
+        self.assertNotEquals(
+            selector_adapter_languages, selector_viewlet_languages
+        )
 
     def assertFullyTranslatedPages(self):
         a = createContentInContainer(
-            self.portal['en'], 'Document', title=u"Test document")
+            self.portal['en'], 'Document', title=u"Test document"
+        )
 
         a_ca = api.translate(a, 'ca')
         a_ca.setTitle(u"Test Document (CA)")
@@ -98,74 +108,87 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         notify(ObjectModifiedEvent(a_ca))
         notify(ObjectModifiedEvent(a_es))
 
-        selector_viewlet =\
-            LanguageSelectorViewlet(a, self.request, None, None)
+        selector_viewlet = LanguageSelectorViewlet(a, self.request, None, None)
         selector_viewlet.update()
         selector_viewlet_languages = selector_viewlet.languages()
 
-        self.assertEqual(selector_viewlet_languages, [{
-            'code': u'en',
-            u'flag': u'/++resource++country-flags/gb.gif',
-            u'name': u'English',
-            u'native': u'English',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal_url,
-                'tg': ITG(a),
-                'lang': 'en',
-                'query': '?set_language=en'
-            },
-            'selected': True,
-            'translated': True,
-        }, {
-            'code': u'ca',
-            u'flag': u'/++resource++language-flags/ca.gif',
-            u'name': u'Catalan',
-            u'native': u'Catal\xe0',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal_url,
-                'tg': ITG(a),
-                'lang': 'ca',
-                'query': '?set_language=ca'
-            },
-            'selected': False,
-            'translated': True,
-        }, {
-            'code': u'es',
-            u'flag': u'/++resource++country-flags/es.gif',
-            u'name': u'Spanish',
-            u'native': u'Espa\xf1ol',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal_url,
-                'tg': ITG(a),
-                'lang': 'es',
-                'query': '?set_language=es'
-            },
-            'selected': False,
-            'translated': True,
-        }])
+        self.assertEqual(
+            selector_viewlet_languages,
+            [
+                {
+                    'code': u'en',
+                    u'flag': u'/++resource++country-flags/gb.gif',
+                    u'name': u'English',
+                    u'native': u'English',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal_url,
+                        'tg': ITG(a),
+                        'lang': 'en',
+                        'query': '?set_language=en',
+                    },
+                    'selected': True,
+                    'translated': True,
+                },
+                {
+                    'code': u'ca',
+                    u'flag': u'/++resource++language-flags/ca.gif',
+                    u'name': u'Catalan',
+                    u'native': u'Catal\xe0',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal_url,
+                        'tg': ITG(a),
+                        'lang': 'ca',
+                        'query': '?set_language=ca',
+                    },
+                    'selected': False,
+                    'translated': True,
+                },
+                {
+                    'code': u'es',
+                    u'flag': u'/++resource++country-flags/es.gif',
+                    u'name': u'Spanish',
+                    u'native': u'Espa\xf1ol',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal_url,
+                        'tg': ITG(a),
+                        'lang': 'es',
+                        'query': '?set_language=es',
+                    },
+                    'selected': False,
+                    'translated': True,
+                },
+            ],
+        )
 
         transaction.commit()
 
         self.browser.open(selector_viewlet_languages[0]['url'])
-        self.assertEqual(self.browser.url,
-                         a.absolute_url() + '?set_language=en')
+        self.assertEqual(
+            self.browser.url, a.absolute_url() + '?set_language=en'
+        )
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         self.browser.open(selector_viewlet_languages[1]['url'])
-        self.assertEqual(self.browser.url,
-                         a_ca.absolute_url() + '?set_language=ca')
+        self.assertEqual(
+            self.browser.url, a_ca.absolute_url() + '?set_language=ca'
+        )
 
-        self.assertIn(u'lang="ca"'.encode("utf-8"), self.browser.contents)
+        self.assertIn('lang="ca"', self.browser.contents)
 
         self.browser.open(selector_viewlet_languages[2]['url'])
-        self.assertEqual(self.browser.url,
-                         a_es.absolute_url() + '?set_language=es')
-        self.assertIn(u'lang="es"'.encode("utf-8"), self.browser.contents)
+        self.assertEqual(
+            self.browser.url, a_es.absolute_url() + '?set_language=es'
+        )
+        self.assertIn('lang="es"', self.browser.contents)
 
     def test_languages_fully_translated_by_closest(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         self.assertFullyTranslatedPages()
@@ -173,7 +196,8 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_languages_fully_translated_by_dialog(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         self.assertFullyTranslatedPages()
@@ -182,12 +206,15 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         wftool = getToolByName(self.portal, 'portal_workflow')
 
         f_en = createContentInContainer(
-            self.portal['en'], 'Folder', title=u"Test folder")
+            self.portal['en'], 'Folder', title=u"Test folder"
+        )
         wftool.doActionFor(f_en, 'publish')
 
         a_en = createContentInContainer(
-            self.portal['en']['test-folder'], 'Document',
-            title=u"Test document")
+            self.portal['en']['test-folder'],
+            'Document',
+            title=u"Test document",
+        )
         wftool.doActionFor(a_en, 'publish')
 
         f_ca = api.translate(f_en, 'ca')
@@ -202,17 +229,18 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         self.setUpPartiallyTranslatedContent()
 
         a_ca = api.translate(
-            self.portal['en']['test-folder']['test-document'], 'ca')
+            self.portal['en']['test-folder']['test-document'], 'ca'
+        )
         a_ca.setTitle(u"Test document CA")
         wftool.doActionFor(a_ca, 'publish')
 
-        f_es = api.translate(
-            self.portal['en']['test-folder'], 'es')
+        f_es = api.translate(self.portal['en']['test-folder'], 'es')
         f_es.setTitle(u"Test folder ES")
         wftool.doActionFor(f_es, 'publish')
 
         a_es = api.translate(
-            self.portal['en']['test-folder']['test-document'], 'es')
+            self.portal['en']['test-folder']['test-document'], 'es'
+        )
         a_es.setTitle(u"Test document ES")
         wftool.doActionFor(a_es, 'publish')
 
@@ -222,13 +250,15 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         # Define selector policy
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         wftool = getToolByName(self.portal, "portal_workflow")
 
         a_ca = createContentInContainer(
-            self.portal['ca'], 'Document', title=u"Test document")
+            self.portal['ca'], 'Document', title=u"Test document"
+        )
         a_ca.setLanguage('ca')
         wftool.doActionFor(a_ca, 'publish')
 
@@ -258,7 +288,8 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_languages_partially_translated_by_closest(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         # Setup content
@@ -274,10 +305,9 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         # Check that EN translation is found
         self.browser.open(selector_languages[0]['url'])
         self.assertEqual(
-            self.browser.url,
-            a_en.absolute_url() + '?set_language=en'
+            self.browser.url, a_en.absolute_url() + '?set_language=en'
         )
-        self.assertIn(u'lang="en"'.encode("utf-8"), self.browser.contents)
+        self.assertIn('lang="en"', self.browser.contents)
         # But extra check, because English is the default?
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
@@ -285,23 +315,23 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         # is found
         self.browser.open(selector_languages[1]['url'])
         self.assertEqual(
-            self.browser.url,
-            f_ca.absolute_url() + '?set_language=ca'
+            self.browser.url, f_ca.absolute_url() + '?set_language=ca'
         )
-        self.assertIn(u'lang="ca"'.encode("utf-8"), self.browser.contents)
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check that ES translation is missing and only the LRF is found
         self.browser.open(selector_languages[2]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal['es'].absolute_url() + '?set_language=es'
+            self.portal['es'].absolute_url() + '?set_language=es',
         )
-        self.assertIn(u'lang="es"'.encode("utf-8"), self.browser.contents)
+        self.assertIn('lang="es"', self.browser.contents)
 
     def test_languages_partially_translated_by_dialog(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         # Setup content
@@ -316,11 +346,10 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         # Check that EN translation is found
         self.browser.open(selector_languages[0]['url'])
         self.assertEqual(
-            self.browser.url,
-            a_en.absolute_url()+'?set_language=en'
+            self.browser.url, a_en.absolute_url() + '?set_language=en'
         )
         # But extra check, because English is the default?
-        self.assertIn(u'lang="en"'.encode("utf-8"), self.browser.contents)
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         tgid = selector_languages[1]['url'].split('/')[-2]
@@ -329,31 +358,30 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         self.browser.open(selector_languages[1]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal.absolute_url() +
-            NOT_TRANSLATED_YET_TEMPLATE +
-            '/' + tgid + '?set_language=ca'
+            self.portal.absolute_url()
+            + NOT_TRANSLATED_YET_TEMPLATE
+            + '/'
+            + tgid
+            + '?set_language=ca',
         )
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check that ES is not yet translated
         self.browser.open(selector_languages[2]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal.absolute_url() +
-            NOT_TRANSLATED_YET_TEMPLATE +
-            '/' + tgid + '?set_language=es'
+            self.portal.absolute_url()
+            + NOT_TRANSLATED_YET_TEMPLATE
+            + '/'
+            + tgid
+            + '?set_language=es',
         )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="es"', self.browser.contents)
 
     def assertSiteRoot(self):
         self.selector_viewlet = LanguageSelectorViewlet(
-            self.portal, self.request, None, None)
+            self.portal, self.request, None, None
+        )
         self.selector_viewlet.update()
         selector_languages = self.selector_viewlet.languages()
 
@@ -362,44 +390,42 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         # Check EN root
         self.assertEqual(
             selector_languages[0]['url'],
-            (self.portal.absolute_url() +
-             '/@@multilingual-selector/notg/en?set_language=en')
+            (
+                self.portal.absolute_url()
+                + '/@@multilingual-selector/notg/en?set_language=en'
+            ),
         )
         self.browser.open(selector_languages[0]['url'])
-        self.assertIn(
-            u'lang="en"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         # Check CA root
         self.assertEqual(
             selector_languages[1]['url'],
-            (self.portal.absolute_url() +
-             '/@@multilingual-selector/notg/ca?set_language=ca')
+            (
+                self.portal.absolute_url()
+                + '/@@multilingual-selector/notg/ca?set_language=ca'
+            ),
         )
         self.browser.open(selector_languages[1]['url'])
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check ES root
         self.browser.open(selector_languages[2]['url'])
         self.assertEqual(
             selector_languages[2]['url'],
-            (self.portal.absolute_url() +
-             '/@@multilingual-selector/notg/es?set_language=es')
+            (
+                self.portal.absolute_url()
+                + '/@@multilingual-selector/notg/es?set_language=es'
+            ),
         )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="es"', self.browser.contents)
 
     def test_siteroot_by_closest(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         self.assertSiteRoot()
@@ -407,98 +433,102 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_siteroot_by_dialog(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         self.assertSiteRoot()
 
     def assertRootFolders(self):
         self.selector_viewlet = LanguageSelectorViewlet(
-            self.portal['en'], self.request, None, None)
+            self.portal['en'], self.request, None, None
+        )
         self.selector_viewlet.update()
 
         selector_languages = self.selector_viewlet.languages()
         tg = ITG(self.portal['en'])
 
-        self.assertEqual(selector_languages, [{
-            'code': u'en',
-            u'flag': u'/++resource++country-flags/gb.gif',
-            u'name': u'English',
-            u'native': u'English',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal.absolute_url(),
-                'tg': tg,
-                'lang': 'en',
-                'query': '?set_language=en'
-            },
-            'selected': True,
-            'translated': True,
-        }, {
-            'code': u'ca',
-            u'flag': u'/++resource++language-flags/ca.gif',
-            u'name': u'Catalan',
-            u'native': u'Catal\xe0',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal.absolute_url(),
-                'tg': tg,
-                'lang': 'ca',
-                'query': '?set_language=ca'
-            },
-            'selected': False,
-            'translated': True,
-        }, {
-            'code': u'es',
-            u'flag': u'/++resource++country-flags/es.gif',
-            u'name': u'Spanish',
-            u'native': u'Espa\xf1ol',
-            'url': SELECTOR_VIEW_TEMPLATE % {
-                'url': self.portal.absolute_url(),
-                'tg': tg,
-                'lang': 'es',
-                'query': '?set_language=es'
-            },
-            'selected': False,
-            'translated': True,
-        }])
+        self.assertEqual(
+            selector_languages,
+            [
+                {
+                    'code': u'en',
+                    u'flag': u'/++resource++country-flags/gb.gif',
+                    u'name': u'English',
+                    u'native': u'English',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal.absolute_url(),
+                        'tg': tg,
+                        'lang': 'en',
+                        'query': '?set_language=en',
+                    },
+                    'selected': True,
+                    'translated': True,
+                },
+                {
+                    'code': u'ca',
+                    u'flag': u'/++resource++language-flags/ca.gif',
+                    u'name': u'Catalan',
+                    u'native': u'Catal\xe0',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal.absolute_url(),
+                        'tg': tg,
+                        'lang': 'ca',
+                        'query': '?set_language=ca',
+                    },
+                    'selected': False,
+                    'translated': True,
+                },
+                {
+                    'code': u'es',
+                    u'flag': u'/++resource++country-flags/es.gif',
+                    u'name': u'Spanish',
+                    u'native': u'Espa\xf1ol',
+                    'url': SELECTOR_VIEW_TEMPLATE
+                    % {
+                        'url': self.portal.absolute_url(),
+                        'tg': tg,
+                        'lang': 'es',
+                        'query': '?set_language=es',
+                    },
+                    'selected': False,
+                    'translated': True,
+                },
+            ],
+        )
 
         # Check EN
         self.browser.open(selector_languages[0]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal['en'].absolute_url()+'?set_language=en'
+            self.portal['en'].absolute_url() + '?set_language=en',
         )
-        self.assertIn(
-            u'lang="en"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         # Check CA
         self.browser.open(selector_languages[1]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal.ca.absolute_url()+'?set_language=ca'
+            self.portal.ca.absolute_url() + '?set_language=ca',
         )
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check ES
         self.browser.open(selector_languages[2]['url'])
         self.assertEqual(
             self.browser.url,
-            self.portal.es.absolute_url()+'?set_language=es'
+            self.portal.es.absolute_url() + '?set_language=es',
         )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="es"', self.browser.contents)
 
     def test_languages_root_folders_by_dialog(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         self.assertRootFolders()
@@ -506,7 +536,8 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_languages_root_folders_by_closest(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         self.assertRootFolders()
@@ -514,7 +545,8 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_languages_preserve_view(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         self.setUpPartiallyTranslatedContent()
@@ -535,41 +567,33 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         self.browser.open(selector_languages[0]['url'])
         self.assertEqual(
             self.browser.url,
-            f_en.absolute_url()+'/contact-info?set_language=en'
+            f_en.absolute_url() + '/contact-info?set_language=en',
         )
-        self.assertIn(
-            u'lang="en"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         # Check CA
         self.browser.open(selector_languages[1]['url'])
         self.assertEqual(
             self.browser.url,
-            f_ca.absolute_url()+'/contact-info?set_language=ca'
+            f_ca.absolute_url() + '/contact-info?set_language=ca',
         )
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check ES
         self.browser.open(selector_languages[2]['url'])
         # Here @@search isn't preserved because we've gone up a level
         self.assertEqual(
             self.browser.url,
-            self.portal.es.absolute_url()+'?set_language=es'
+            self.portal.es.absolute_url() + '?set_language=es',
         )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="es"', self.browser.contents)
 
         # Shouldn't do for the not-translated-yet thingie when dialog is on
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         selector = LanguageSelectorViewlet(f_en, self.request, None, None)
@@ -582,24 +606,18 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         self.browser.open(selector_languages[0]['url'])
         self.assertEqual(
             self.browser.url,
-            f_en.absolute_url()+'/contact-info?set_language=en'
+            f_en.absolute_url() + '/contact-info?set_language=en',
         )
-        self.assertIn(
-            u'lang="en"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         # Check CA
         self.browser.open(selector_languages[1]['url'])
         self.assertEqual(
             self.browser.url,
-            f_ca.absolute_url()+'/contact-info?set_language=ca'
+            f_ca.absolute_url() + '/contact-info?set_language=ca',
         )
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check ES
         self.browser.open(selector_languages[2]['url'])
@@ -607,14 +625,13 @@ class TestLanguageSelectorBasics(unittest.TestCase):
         tgid = selector_languages[2]['url'].split('/')[-3]
         self.assertEqual(
             self.browser.url,
-            self.portal.absolute_url() +
-            NOT_TRANSLATED_YET_TEMPLATE +
-            '/' + tgid + '?set_language=es'
+            self.portal.absolute_url()
+            + NOT_TRANSLATED_YET_TEMPLATE
+            + '/'
+            + tgid
+            + '?set_language=es',
         )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="es"', self.browser.contents)
 
     def assertLanguagesPreserveQuery(self, policy):
 
@@ -632,56 +649,58 @@ class TestLanguageSelectorBasics(unittest.TestCase):
 
         tgid = selector_languages[2]['url'].split('/')[-3]
         untranslated_url = {
-            'closest': (self.portal['es'].absolute_url() +
-                        '?int=1&uni=pres%C3%98rved&set_language=es'),
-            'dialog': (self.portal.absolute_url() +
-                       NOT_TRANSLATED_YET_TEMPLATE + '/' + tgid +
-                       '?int=1&uni=pres%C3%98rved&set_language=es')
+            'closest': (
+                self.portal['es'].absolute_url()
+                + '?int=1&uni=pres%C3%98rved&set_language=es'
+            ),
+            'dialog': (
+                self.portal.absolute_url()
+                + NOT_TRANSLATED_YET_TEMPLATE
+                + '/'
+                + tgid
+                + '?int=1&uni=pres%C3%98rved&set_language=es'
+            ),
         }
 
         transaction.commit()
 
         # Check EN
         self.browser.open(selector_languages[0]['url'])
-        self.assertEqual(
+        self.assertUrlsEqual(
             self.browser.url,
-            (f_en.absolute_url() +
-             '/contact-info?int=1&uni=pres%C3%98rved&set_language=en')
+            (
+                f_en.absolute_url()
+                + '/contact-info?int=1&uni=pres%C3%98rved&set_language=en'
+            ),
         )
-        self.assertIn(
-            u'lang="en"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="en"', self.browser.contents)
         self.assertRegexpMatches(self.browser.contents, r"You\s*are here")
 
         # Check CA
         self.browser.open(selector_languages[1]['url'])
-        self.assertEqual(
+        self.assertUrlsEqual(
             self.browser.url,
-            (f_ca.absolute_url() +
-             '/contact-info?int=1&uni=pres%C3%98rved&set_language=ca')
+            (
+                f_ca.absolute_url()
+                + '/contact-info?int=1&uni=pres%C3%98rved&set_language=ca'
+            ),
         )
-        self.assertIn(
-            u'lang="ca"'.encode("utf-8"),
-            self.browser.contents
-        )
+        self.assertIn('lang="ca"', self.browser.contents)
 
         # Check ES
         self.browser.open(selector_languages[2]['url'])
         # Here @@search isn't preserved because we've got the dialog
-        self.assertEqual(
-            self.browser.url,
-            untranslated_url[policy]
-        )
-        self.assertIn(
-            u'lang="es"'.encode("utf-8"),
-            self.browser.contents
-        )
+        url = self.browser.url
+        if six.PY2 and isinstance(url, six.binary_type):
+            url = url.decode('utf8')
+        self.assertUrlsEqual(url, untranslated_url[policy])
+        self.assertIn('lang="es"', self.browser.contents)
 
     def test_languages_preserve_by_closest(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'closest'
 
         self.assertLanguagesPreserveQuery('closest')
@@ -689,7 +708,8 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     def test_languages_preserve_by_dialog(self):
         self.registry = getUtility(IRegistry)
         self.settings = self.registry.forInterface(
-            IMultiLanguageExtraOptionsSchema, prefix="plone")
+            IMultiLanguageExtraOptionsSchema, prefix="plone"
+        )
         self.settings.selector_lookup_translations_policy = 'dialog'
 
         self.assertLanguagesPreserveQuery('dialog')
@@ -755,6 +775,20 @@ class TestLanguageSelectorBasics(unittest.TestCase):
     #          'url': base + 'pres%C3%98rved&set_language=no'}]
     #     self.assertEqual(selector.languages(), expected)
 
+    def assertUrlsEqual(self, url1, url2):
+        """The order of query-strings is sometimes random in python 3
+        This compares urls disregarding the order.
+        """
+        parsed_url_1 = urlparse(safe_nativestring(url1))
+        parse_qs_1 = parse_qs(parsed_url_1.query)
+        parsed_url_2 = urlparse(safe_nativestring(url2))
+        parse_qs_2 = parse_qs(parsed_url_2.query)
+        self.assertEqual(parsed_url_1[0], parsed_url_2[0])
+        self.assertEqual(parsed_url_1[1], parsed_url_2[1])
+        self.assertEqual(parsed_url_1[2], parsed_url_2[2])
+        self.assertEqual(parsed_url_1[3], parsed_url_2[3])
+        self.assertEqual(parse_qs_1, parse_qs_2)
+
 
 class TestLanguageSelectorPostPath(unittest.TestCase):
 
@@ -764,73 +798,62 @@ class TestLanguageSelectorPostPath(unittest.TestCase):
         alsoProvides(self.layer['request'], IPloneAppMultilingualInstalled)
         self.portal = self.layer['portal']
         self.folder = createContentInContainer(
-            self.portal['ca'], 'Folder', title=u"Folder")
+            self.portal['ca'], 'Folder', title=u"Folder"
+        )
         self.document = createContentInContainer(
-            self.portal['ca'], 'Document', title=u"Document")
+            self.portal['ca'], 'Document', title=u"Document"
+        )
         self.request = self.layer['request']
 
     def test_findpath(self):
         self.request['PATH_INFO'] = '/plone/ca/folder/document/whatever'
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/whatever'
-        )
+        self.assertEqual(getPostPath(self.document, self.request), '/whatever')
 
     def test_findpath_match(self):
         self.request['PATH_INFO'] = '/plone/ca/folder/document'
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            ''
-        )
+        self.assertEqual(getPostPath(self.document, self.request), '')
 
     def test_findpath_match_slash(self):
         self.request['PATH_INFO'] = '/plone/ca/folder/document/'
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            ''
-        )
+        self.assertEqual(getPostPath(self.document, self.request), '')
 
     def test_findpath_template(self):
-        self.request['PATH_INFO'] = ('/plone/ca/folder/document/'
-                                     'whatever/foo/edit')
+        self.request['PATH_INFO'] = (
+            '/plone/ca/folder/document/' 'whatever/foo/edit'
+        )
         self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/whatever/foo/edit'
+            getPostPath(self.document, self.request), '/whatever/foo/edit'
         )
 
     def test_findpath_view(self):
         self.request['PATH_INFO'] = '/plone/ca/folder/document/contact-info'
         self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/contact-info'
+            getPostPath(self.document, self.request), '/contact-info'
         )
 
     def test_findpath_vhr(self):
-        self.request['PATH_INFO'] = ('/VirtualHostBase/http/127.0.0.1'
-                                     '/plone/ca/folder/document/'
-                                     'VirtualHostRoot/whatever')
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/whatever'
+        self.request['PATH_INFO'] = (
+            '/VirtualHostBase/http/127.0.0.1'
+            '/plone/ca/folder/document/'
+            'VirtualHostRoot/whatever'
         )
+        self.assertEqual(getPostPath(self.document, self.request), '/whatever')
 
     def test_findpath_vh_marker(self):
-        self.request['PATH_INFO'] = ('/VirtualHostBase/http/127.0.0.1'
-                                     '/plone/ca/folder/document/'
-                                     'VirtualHostRoot/_vh_foo/whatever')
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/whatever'
+        self.request['PATH_INFO'] = (
+            '/VirtualHostBase/http/127.0.0.1'
+            '/plone/ca/folder/document/'
+            'VirtualHostRoot/_vh_foo/whatever'
         )
+        self.assertEqual(getPostPath(self.document, self.request), '/whatever')
 
     def test_findpath_vhr_and_traverser(self):
-        self.request['PATH_INFO'] = ('/VirtualHostBase/http/127.0.0.1'
-                                     '/plone/ca/folder/document/++theme++foo/'
-                                     'VirtualHostRoot/whatever')
-        self.assertEqual(
-            getPostPath(self.document, self.request),
-            '/whatever'
+        self.request['PATH_INFO'] = (
+            '/VirtualHostBase/http/127.0.0.1'
+            '/plone/ca/folder/document/++theme++foo/'
+            'VirtualHostRoot/whatever'
         )
+        self.assertEqual(getPostPath(self.document, self.request), '/whatever')
 
 
 class TestLanguageSelectorAddQuery(unittest.TestCase):
@@ -845,55 +868,65 @@ class TestLanguageSelectorAddQuery(unittest.TestCase):
     def test_formvariables(self):
         self.request.form['one'] = 1
         self.request.form['two'] = 2
-        self.assertEqual(
-            addQuery(self.request, self.url),
-            self.url+'?two:int=2&one:int=1'
+        self.assertUrlsEqual(
+            addQuery(self.request, self.url), self.url + '?two:int=2&one:int=1'
         )
 
     def test_formvariables_sequences(self):
-        self.request.form['one'] = ('a', )
+        self.request.form['one'] = ('a',)
         self.request.form['two'] = ['b', 2]
-        self.assertEqual(
+        self.assertUrlsEqual(
             addQuery(self.request, self.url),
-            self.url+'?two:list=b&two:int:list=2&one=%28%27a%27%2C%29'
+            self.url + '?two:list=b&two:int:list=2&one=%28%27a%27%2C%29',
         )
 
     def test_formvariables_unicode(self):
         self.request.form['one'] = u'Før'
         self.request.form['two'] = u'foo'
-        self.assertEqual(
+        self.assertUrlsEqual(
             addQuery(self.request, self.url),
-            self.url+'?two=foo&one=F%C3%B8r'
+            self.url + '?two=foo&one=F%C3%B8r',
         )
 
     def test_formvariables_utf8(self):
-        self.request.form['one'] = u'Før'.encode("utf-8")
+        self.request.form['one'] = u'Før'
         self.request.form['two'] = u'foo'
-        self.assertEqual(
+        self.assertUrlsEqual(
             addQuery(self.request, self.url),
-            self.url+'?two=foo&one=F%C3%B8r'
+            self.url + '?two=foo&one=F%C3%B8r',
         )
 
     def test_formvariables_object(self):
         self.request.form['one'] = '1'
         self.request.form['two'] = EvilObject()
-        self.assertEqual(
-            addQuery(self.request, self.url),
-            self.url
-        )
+        self.assertUrlsEqual(addQuery(self.request, self.url), self.url)
 
     def test_formvariables_exclude(self):
         self.request.form['one'] = 1
         self.request.form['two'] = 2
-        self.assertEqual(
+        self.assertUrlsEqual(
             addQuery(self.request, self.url, exclude=('two',)),
-            self.url+'?one:int=1'
+            self.url + '?one:int=1',
         )
 
     def test_formvariables_extras(self):
         self.request.form['one'] = 1
         self.request.form['two'] = 2
-        self.assertEqual(
+        self.assertUrlsEqual(
             addQuery(self.request, self.url, three=3),
-            self.url+'?one:int=1&three:int=3&two:int=2'
+            self.url + '?one:int=1&three:int=3&two:int=2',
         )
+
+    def assertUrlsEqual(self, url1, url2):
+        """The order of query-strings is sometimes random in python 3
+        This compares urls disregarding the order.
+        """
+        parsed_url_1 = urlparse(url1)
+        parse_qs_1 = parse_qs(parsed_url_1.query)
+        parsed_url_2 = urlparse(url2)
+        parse_qs_2 = parse_qs(parsed_url_2.query)
+        self.assertEqual(parsed_url_1[0], parsed_url_2[0])
+        self.assertEqual(parsed_url_1[1], parsed_url_2[1])
+        self.assertEqual(parsed_url_1[2], parsed_url_2[2])
+        self.assertEqual(parsed_url_1[3], parsed_url_2[3])
+        self.assertEqual(parse_qs_1, parse_qs_2)