+++ CHANGES.rst
@@ -34,6 +34,9 @@ New features:
 - Make plone.app.controlpanel optional (no longer there in Plone 5.2).
   [jensens]
 
+- Add compatability with python 3
+  [pbauer]
+
 Bug fixes:
 
 - Upgrade step to profile version 3 was lost and now recreated.
@@ -46,6 +49,9 @@ Bug fixes:
 - Don't fail, if multilingual selector is called without query
   [tomgross]
 
+- Fix tests for py3
+  [pbauer]
+
 - Fix connecting of documents
   [tomgross]
 
@@ -55,7 +61,8 @@ Bug fixes:
 
 Bug fixes:
 
-- Fixed tests now that Catalan has translated ‘assets’ into ‘recursos’. [maurits]
+- Fixed tests now that Catalan has translated ‘assets’ into ‘recursos’. 
+  [maurits]
 
 
 5.2.0 (2018-04-04)